Miscellaneous stability improvements.
- introduce the archiverTimeout parameter.

The archiverTimeout causes the routine that archives deltas for a table to
close and archive the current file if it receives no changes since the last
one for the duration of the timeout.

Previously, a particular routine of the logical backup called `closeOldFiles`
was responsible for closing all open files after the hard-coded interval of 3h
(that value now servers as a default for the `archiverTimeout`).
Unfortunately, it was doing it in a way that was not concurrency-safe, causing
errors on subsequent writes to that closed file. The tentative fix was to
check if the file was still “active”; this patch implements a more long-term
one of archiving such files on a timeout by the archiver routine. The archiver
employs a new `DeltaFileAccessMutex` to avoid archiving a file that the delta
handler writes.

- Make sure the archiver doesn’t block senders to the  `archiveFiles` channel
under load by splitting the archiver code into two routines; the listener one
that puts tasks into the queue, and the worker routine that reads from the
queue and archives files accordingly.

The changes above related to the archiverTimeout added some extra work to the
archiver, and by design, the archiver shouldn't execute any non-trivial tasks
synchronously to be able to read the next filename from the channel without
blocking writers. The changes split the archiver into two parts: one listens
to a channel, puts the filenames into the queue and occasionally (every 10
minutes) wakes up to check if it should close the current file due to the
timeout; the worker part reads delta filenames from the queue and does the
heavy lifting of moving them to the final directory. That split also allows to
handle the archiverTimeout nicely: the main routine closes the file and puts
its name to the queue to archive it later. Also, that allows retrying the
archiving action several times if an error occurs without blocking the writes
to the `archiveFiles` channel.

- Close a basebackup transaction when trying to backup a table with no rows.

Previously, the code bailed up without any cleanup, leaving the dump
transaction open and preventing any further backups on the corresponding
tables.

- Prevent archiving deltas that precede the latest backups.

The original basebackup code removed those deltas; however, at leat two issues
were present:

The deletion code only considered deltas the final directory, leaving some
files in the temp intact; those deltas would be archived in a later point in
time, making its way to the final directory despite our promise to get rid of
them.

The dump had acquired the WAL log sequence number (lsn), used as a cutoff
point, at the beginning of the dump transaction. That cause deltas for
transactions started before the dump but not committed until after the dump
start to be removed, leading to a data loss. The new code takes the lsn of the
delta file before and after the starting point of the dump and sets the latest
of both that precedes the dump as a cutoff point (usually both should be the
same, unless there is a race condition where the delta is rotated during the
time the dump finds its lsn).

- Avoid queuing multiple backups for a single table.

Introduce IsBasebackupPending, SetBasebackupPending and ClearBasebackupPending
methods on a to be used when the backup is queued and dequeued to avoid
queuing the same backup multiple times.