Minor improvements around the initialization code and backups. (#12)
* Minor improvements around the initialization code and backups.

Remove “tables” configuration parameter. The list of tables to replicate can be
fully controlled by the publication.

Unify the code that checks whether tables to replicate have a replica identity
and the one that adds the tables to the list of those we replica. Previously,
the code that checked for the replica identity was not aware of publications.

Drop the logical replication slot that we use for fetching lsns.

Minor improvements in debug message verbosity and SQL queries.

* Quote the SQL argument, fix some comment, use log instead of fmt.
* Some minor changes around passing and quoting parameters.