Get rid of the archive channel in favor of the queue
Queue files to archive instead of sending them to the buffered channel. Since archiver itself was using the queue to pass the files from the channel to the actual worker goroutine, the changes here just eliminate the extra channel in favor of passing those files directly to the queue.

Split archiver into the archiver (archive files) and janitor (closes and truncates untouched files open for more than 3h) goroutines.