Write current deltas into a memory buffer instead of a disk file.
Flush the buffer into a disk file once we have filled the delta up to a
threshold, or when the archiver decides to archive the file due to the
timeout.

Call fsync on the files and parent directory entries if fsync is on.
This is done both when the buffers are flushed, as well as when the file
is copied to the final directory.