Fix os.Stat call in the archiver.
Previously, we've checked for os.isExists, which is wrong, since
stat will just return nil as an error in that case.

That resulted in the condition of the destination file being not empty
to be ignored completely. Another fix would address that.

Use the uint64 representation of the LSN to form a delta dir name, as
the natvie one will produce a "A/B" string.