Remove the internal relation maps from the logicalBackup.
Their only purpose was to catch the cases of table rename or drop and
recreate when we see the relation message; however, by switching to
OIDs we would eliminate the need to handle those cases completely.

We can just rely on the backupTables map to track new tables and, in
perspective, let the basebackup decide to exclude a table from the
backup if the table is not there anymore.