Simplify backup table path (#19)
* use Oids instead of names in path
* Fix for reading the oid type from the DB and getting path from the oid.
* Fix os.Stat call in the archiver.

Previously, we've checked for os.isExists, which is wrong, since
stat will just return nil as an error in that case.

That resulted in the condition of the destination file being not empty
to be ignored completely. Another fix would address that.

* Fix os.Stat call in the archiver.

Previously, we've checked for os.isExists, which is wrong, since
stat will just return nil as an error in that case.

That resulted in the condition of the destination file being not empty
to be ignored completely. Another fix would address that.

Use the uint64 representation of the LSN to form a delta dir name, as
the natvie one will produce a "A/B" string.

* Overwrite backup and info files.

Overwrite files that are supposed to be only one instance of (basebackup
and info) instead of copying them. Don't do this for deltas, though, as
there is an opportunity to lose data by overwriting the wrong delta.

* Remove the internal relation maps from the logicalBackup.

Their only purpose was to catch the cases of table rename or drop and
recreate when we see the relation message; however, by switching to
OIDs we would eliminate the need to handle those cases completely.

We can just rely on the backupTables map to track new tables and, in
perspective, let the basebackup decide to exclude a table from the
backup if the table is not there anymore.