bugfixes (#33)
* Set prometheus port to 1999 when not defined explicitly.
* Safeguard against effects of dropping a table included in the backup.

Make sure dropped tables are excluded from the backup as soon as we
notice they are not present. The right place for this seems to be the
locking routine called from RunBasebackup, as base backup is the only
place we could interact with those and lock comes as a first SQL command
(we need to lock the table even before fething its structure in order to
protect against concurrent modifications/deletes).

Also set prometheus port to 1999 when not defined explicitly.

* Protect BackupTables structure with a mutex.

We need to lock when deleting or adding new tables concurrently with
reading existing ones. SyncMap would be an alternative, although it uses
interfaces to be generic and therefore relaxes the type checks.