Report current LSN as a flush LSN during the table segment switch (#30)
* Set flush LSN during the flush of the segment file
* Send flush LSN to the server on commit if it has changed.
* Report current LSN as a flush LSN during the table segment switch.

Expose currentLSN from the CopyData message (repMsg.WalMessage) and use
it during the segment switch to set flush LSN too.

Report current LSN instead of the previous transaction commit LSN as a
flush LSN when the segment file is flushed.

Use the minimum of all flush LSNs for tables touched in the current
transaction to determine the global flush LSN to report to the server
and report it on commit if it is different from the previous one.

Also report current flush LSN at start, when reading the state file.

Rename currentTransactionLSN to the TransactionCommitLSN.

Remove storedFlushLSN and StartLSN, as they are not used anymore. Instead,
rely on FlushLSN to replace both. Also remove LastWrittenCommitLSN, as it
we never need the LSN of the commit not corresponding to the latest
transaction coming from the slot.

Simplify the code around InitSlot by returning InvalidLSN if the slot
doesn’t exist (an existing one will always have a valid FlushLSN).

Add print function for messages, currently unused but potentially helpful
for figuring out which excact messages are coming from the slot.

* Remove some spurious newlines and sendStatusOnCommit parameter.

sendStatusOnCommit is outdated; apart from sending the status every 10s
and on-demand, when the server asks for it, the status is sent
automatically when flushLSN changes.