Persist per-table state between restart. (#36)
* Persist per-table state between restart.

Store and restore per-table indicators between restart. Without it,
base backups  would not be done on time and some deltas could be written
twice. The state file has a table OID in its name, so that we could
locate it easily.

Make sure deltas that are prior to flush point and discarded without
processing. This is normal, since the slot's restart LSN is advanced
some time after the flush LSN.

Rename TableInfoFilename to the BaseBackupStateFileName, as it is only
relevant for the base backup.

Set initialBasebackup to false by default to avoid triggering basebackup
every time the tool is restarted.