Skip already written per-table deltas, fixes for ad6c9f2.
Avoid writing deltas that has already being flushed according to the
per-table flush LSN. The reason we may try to write those is because
the 'common' flush LSN is the minumum of all per-table ones, so for some
tables further ahead we may receive the changes again after the restart.

Fix trying to send a status message before initializing the replication
connection and using an outdated LSN when considering to skip current
message.

Some minor refactoring.