Revert printing of arguments to the bunch of logger calls.
Naive approaches of putting them in a slice or a map before printing
haven't produced either shorter or more readable code, partially because
incapsulating a check into a ternary-operator like function is not
possible with go, as the language evaluations function arguments at call
time. We may revisit this code one day to fetch the description from the
Config struct using reflection and to separate validation of arguments
from the actual printing.