Add zap logger infrastructure.
Having a structure in the logs helps to highlight important data
(like errors or LSNs), especially when importing to the external log
storage.

Another long-standing addition is debug levels, allowing to distinguish
between debug and info messages.

This implementation wraps zap around our own logger module. The module
does not define an interface, as there seem no need for it, although it
would be trivial to add when necessary (i.e. if we consider plugging
another logger or use the same set of functions to pass structured error
messages up to the calling functions).

The changes also include new configuration options for backup and
restore to setup logging level, whether to show code locations in logs
and whether to enable development logging (output to console instead of
json, debug level by default and some other changes set in
zap.NewDevelopmentConfig()).

The changes does not convert existing logging solution to the new one.
This would be done after the changes in
https://github.com/ikitiki/logical_backup/commits/fixes are committed.